refractor de Favorites controller con tests pasados
