+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/FavoritesController.java
@@ -29,6 +29,7 @@ public class FavoritesController {
     @NonNull
     private final JwtServiceImpl jwtService;
 
+
     @NonNull
     private final IChallengeService challengeService;
